Attach/DeAttach Functions
Finished the attach and de-attach functions, building can be attached to tiles if a building has be selected to be placed.
On the other hand, builldings can be de-attached when deletion mode is on.
If neither of these are true (no building to be placed or deletion mode is off), nothing happens as expected
Once a building is de-attached, its object is also destroyed from the scene.
Updated BuyMenu to feature a destroy button inseatd of using 'PageUpButton'
Re-named 'SetDelectionMode' to 'SetDeletionMode'